kuiserverv2jobtracker: Don't terminate a null JobView
There's no point. It can probably happen that another update call
added back the view struct but we didn't have a jobview, so it would
eventually crash when e.g. plasmashell restarted.

+++ src/kuiserverv2jobtracker.cpp
@@ -214,7 +214,9 @@ void KUiServerV2JobTracker::registerJob(KJob *job)
                     const uint errorCode = oldState.value(QStringLiteral("errorCode")).toUInt();
                     const QString errorMessage = oldState.value(QStringLiteral("errorMessage")).toString();
 
-                    view.jobView->terminate(errorCode, errorMessage, QVariantMap() /*hints*/);
+                    if (view.jobView) {
+                        view.jobView->terminate(errorCode, errorMessage, QVariantMap() /*hints*/);
+                    }
 
                     delete view.jobView;
                     d->jobViews.remove(job);
